requirements.txt
Flask
pandas
requests
python-metar